added auth0 roles functionality to the employee data page
